Project-REST-API-PostgreSQL/requirements.txt
flask
# Auto Restart App, when any change in the code
python-dotenv
# To connect to DataBase
psycopg2-binary